Move prop interfaces to individual files
